+++ bindings/pybind11/tests/test_idyntree_modelio_urdf.py
@@ -1,4 +1,6 @@
 """Tests for idyntree-model-io-urdf Python bindings."""
+import os
+import tempfile
 import unittest
 
 import idyntree.pybind as iDynTree
@@ -51,6 +53,72 @@ class IDynTreeModelIoUrdfTest(unittest.TestCase):
     exporter.init(model, sensors=sensors)
     self.assertEqual(
         exporter.sensors.get_nr_of_sensors(iDynTree.ACCELEROMETER), 1)
+
+  def test_importer(self):
+    model = (r'<robot name="test_robot">'
+             r'<link name="link1" />'
+             r'<link name="link2" />'
+             r'<link name="link3" />'
+             r'<link name="link4" />'
+             r'<joint name="joint1" type="continuous">'
+             r'<parent link="link1"/>'
+             r'<child link="link2"/>'
+             r'</joint>'
+             r'<joint name="joint2" type="continuous">'
+             r'<parent link="link1"/>'
+             r'<child link="link3"/>'
+             r'</joint>'
+             r'<joint name="joint3" type="continuous">'
+             r'<parent link="link3"/>'
+             r'<child link="link4"/>'
+             r'</joint>'
+             r'</robot>')
+
+    tmpdir = tempfile.TemporaryDirectory()
+    urdf_filename = os.path.join(tmpdir.name, 'model.urdf')
+    with open(urdf_filename, 'w') as f:
+      f.write(model)
+
+    importer = iDynTree.ModelLoader()
+    importer.load_model_from_file(urdf_filename)
+    self.assertEqual(4, importer.model.get_nr_of_links())
+    self.assertEqual(3, importer.model.get_nr_of_joints())
+
+  def test_importer_reduced_model(self):
+    model = (r'<robot name="test_robot">'
+             r'<link name="link1" />'
+             r'<link name="link2" />'
+             r'<link name="link3" />'
+             r'<link name="link4" />'
+             r'<joint name="joint1" type="continuous">'
+             r'<parent link="link1"/>'
+             r'<child link="link2"/>'
+             r'</joint>'
+             r'<joint name="joint2" type="continuous">'
+             r'<parent link="link1"/>'
+             r'<child link="link3"/>'
+             r'</joint>'
+             r'<joint name="joint3" type="continuous">'
+             r'<parent link="link3"/>'
+             r'<child link="link4"/>'
+             r'</joint>'
+             r'</robot>')
+
+    tmpdir = tempfile.TemporaryDirectory()
+    urdf_filename = os.path.join(tmpdir.name, 'model.urdf')
+    with open(urdf_filename, 'w') as f:
+      f.write(model)
+
+    importer = iDynTree.ModelLoader()
+    importer.load_model_from_file(urdf_filename)
+    importer.load_reduced_model_from_full_model(importer.model,
+                                                ('joint1', 'joint3'))
+    # The reduced model will have the 2 joints we specified, and one less body
+    # as the missing joint made two bodies being merged together.
+    self.assertEqual(3, importer.model.get_nr_of_links())
+    self.assertEqual(2, importer.model.get_nr_of_joints())
+    self.assertEqual('joint1', importer.model.get_joint_name(0))
+    self.assertEqual('joint3', importer.model.get_joint_name(1))
 
 
 if __name__ == "__main__":